backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
pymongo 4.15.5
dnspython 2.8.0

# testing
pytest
pytest-asyncio
httpx
mongomock_motor
